2015-05-22 Roman - SYS: Targets "*-lzma" into Makefile.am fixed for force using LZMA by --use-compress-program=lzma but argument --lzma often uses XZ.     {debian,debian_mod}/changelog, *.spec: Work version scheme naming set to 0.9+rNNNN for adapt to the AutoBuilder, 	allow next update to 0.9.N and the Work version's SVN revision present into the program binary.     {debian,debian_mod}/rules: Autoreconf call by dh_autoreconf restored.     data/openscada_start.in: "subst" command using, as the ALTLinux specific, replaced to generic "sed". - FIX:DB.FireBird: Configure template AX_LIB_FIREBIRD adapted for configure and build that with 64bit libraries, like Mandriva64.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2265 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/protocol/SelfSystem/self.cpp
@@ -104,7 +104,7 @@ int TProt::sesOpen( const string &user, const string &pass, const string &src )
 	    }
 	    ++aI;
 	}
-    if(i_oCnt > singleUserHostLimit() && aOldI != mAuth.end()) {
+    if((int)i_oCnt > singleUserHostLimit() && aOldI != mAuth.end()) {
 	mess_err(nodePath().c_str(), _("Connections from the user '%s' and the source '%s' reached to limit %d. Erasing spare!"),
 	    user.c_str(), TSYS::strLine(src,0).c_str(), singleUserHostLimit());
 	mAuth.erase(aOldI);
